let the user invoke the query
