+++ webapp/pages/topics/main.tsx
@@ -1,7 +1,7 @@
 import { useLocalStorage } from "@mantine/hooks";
 import { Allotment } from "allotment";
 import { useParams } from "react-router-dom";
-import { Topic } from "./topic";
+import { Topic } from "./topic/topic";
 import { TopicList } from "./topic-list";
 
 export const TopicsPage = () => {
